+++ client/plots/geneset.js
@@ -3,23 +3,26 @@ import { showGenesetEdit } from '../dom/genesetEdit.ts'
 import { fillTermWrapper } from '#termsetting'
 import { dofetch3 } from '#common/dofetch'
 
-/*
-FIXME please fully explain functionality and arg, e.g. toolName
-(now that both oncomatrix and gdccluster are both routed through it, how is this connected to the downstream tool)
+// This is a reactive geneset component, meant for use within plotApp.
+//
+// 1. plotApp is initialized with geneset as the first plot/component
+// 2. if there are opts.genes, geneset would simply call its opts.callback right away (not even render the edit UI)
+// 3. if not, then geneset would render the edit UI, and also use its opts.callback to supply the wrapped genes
+// 4. in the opts.callback, the matrix/hierCluster will dispatch plot_delete to remove the geneset as the plot,
+//    and also plot_create to render the target plot (like matrix, hierCluster)
+//
+// The reactivity is rquired to toggle loading overlay visibility
+// and to create a new plot within the same plot app, once a valid geneset is
+// obtained from the server.
+//
 
-This is a reactive geneset component, meant for use with
-plotApp to offer a geneset edit UI when there are no starting gene lst.
-The reactivity has to mostly to do with toggling loading overlay visibility
-and creating a new plot within the same plot app, once a valid geneset is
-obtained from the server.
- - Use dom/genesetEdit to directly use the non-reactive version.
-*/
 class GenesetComp {
 	constructor(opts) {
 		this.type = 'geneset'
 		this.dom = {
 			holder: opts.holder
 		}
+		this.sequenceId = 0
 	}
 
 	getState(appState) {
@@ -33,7 +36,13 @@ class GenesetComp {
 
 	async main() {
 		this.dom.holder.selectAll('*').remove()
+		//const actionSequenceId = this.api.notes('actionSequenceId')
 		const genes = await this.getGenes()
+		// if (this.api.notes('actionSequenceId') !== actionSequenceId) {
+		// 	// (an)other state change(s) has been dispatched between the start and completion of the server request
+		// 	console.warn('aborted state update, the server data corresponds to a stale action.sequenceId')
+		// 	return
+		// }
 		if (!genes?.length) this.render()
 		else
 			setTimeout(async () => {