+++ client/src/launchGdcHierCluster.js
@@ -68,6 +68,7 @@ export async function init(arg, holder, genomes) {
 		if (!Number.isInteger(settings.hierCluster.maxGenes)) settings.hierCluster.maxGenes = 100
 
 		if (arg.filter0 && typeof arg.filter0 != 'object') throw 'arg.filter0 not object'
+		let plot_spliced = false
 
 		const plotAppApi = await appInit({
 			holder: select(arg.holder).select('.sja_root_holder'),
@@ -89,10 +90,9 @@ export async function init(arg, holder, genomes) {
 					}
 				]
 			},
-			app: {
-				features: ['recover'],
-				callbacks: arg.opts?.app?.callbacks || {}
-			},
+			app: arg.opts?.app || {},
+			hierCluster: arg.opts?.hierCluster || {},
+			matrix: arg.opts?.matrix || {},
 			geneset: {
 				mode: 'expression',
 				genome,
@@ -107,6 +107,9 @@ export async function init(arg, holder, genomes) {
 					`)
 				},
 				callback(genesetCompApi, twlst) {
+					// this callback may be called more than once as a user changes a GDC cohort initially,
+					// need to avoid re-deleting/re-creating plots
+					if (plot_spliced || !genesetCompApi) return
 					plotAppApi.dispatch({
 						type: 'plot_splice',
 						subactions: [
@@ -172,13 +175,9 @@ export async function init(arg, holder, genomes) {
 		})
 
 		let hierClusterApi
-		const api = Object.freeze({
+		const api = {
 			type: 'hierCluster',
 			update: async _arg => {
-				if (!plotAppApi) {
-					Object.assign(pendingArg, _arg)
-					return
-				}
 				if ('filter0' in _arg) {
 					plotAppApi.dispatch({
 						type: 'filter_replace',
@@ -193,7 +192,7 @@ export async function init(arg, holder, genomes) {
 					})
 				}
 			}
-		})
+		}
 
 		return api
 	} catch (e) {